refactor: update agent imports and enhance weather tool integration
- Changed import path for mastra in actions.ts for consistency.
- Updated agent name from "assistant" to "assist" in getWeatherInfo function.
- Corrected import statement for getWeatherInfo in form.tsx.
- Integrated weatherTool into the assistant and starterAgent configurations.
- Fixed learning agent reference in starter-agent-tool.ts for clarity.
