+++ app/tests/actions.ts
@@ -1,12 +1,12 @@
 "use server";
- 
-import { mastra } from "../../mastra";
- 
+
+import { mastra } from '@/src/mastra';
+
 export async function getWeatherInfo(formData: FormData) {
   const city = formData.get("city")?.toString();
-  const agent = mastra.getAgent("assistant");
- 
+  const agent = mastra.getAgent("assist");
+
   const result = await agent.generate(`What's the weather like in ${city}?`);
- 
+
   return result.text;
 }
